Create an "erase nop calls" pass and move the std::move erasure there. This makes the erasure of the calls independent of the (particular) code gen.
